% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_icc.R
\name{calculate_theoretical_icc}
\alias{calculate_theoretical_icc}
\title{Calculate Theoretical ICC}
\usage{
calculate_theoretical_icc(sd_u, sd_e)
}
\description{
Computes ICC based on input standard deviations
}
