% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simulation_mlm.R
\name{simulate_mlm_data}
\alias{simulate_mlm_data}
\title{Simulate Multilevel Data}
\usage{
simulate_mlm_data(
  N = 30,
  C = 50,
  g00 = 2,
  g01 = 5,
  tau2_u0 = input$tau2_00,
  tau2_u1 = input$tau2_11,
  cov_uij.u0j = input$tau2_01,
  sigma2 = input$sigma2
)
}
\description{
A function to generate random intercept data.
}
